requirements.txt
asttokens==2.4.1
boto3==1.34.131
botocore==1.34.131
click==8.1.7
decorator==5.1.1
executing==2.0.1
ipython==8.25.0
jedi==0.19.1
jmespath==1.0.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
rich==13.7.1
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
